+++ user-service/src/main/java/com/example/userservice/service/UserService.java
@@ -1,6 +1,6 @@
 package com.example.userservice.service;
 
-import com.example.userservice.entity.User;
+import com.example.userservice.entity.user.User;
 import com.example.userservice.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
